+++ packages/types/src/agent/index.ts
@@ -1,10 +1,11 @@
 import { LLMParams } from 'model-bank';
+
 import { FileItem } from '../files';
 import { KnowledgeBaseItem } from '../knowledgeBase';
 import { FewShots } from '../llm';
 import { LobeAgentChatConfig } from './chatConfig';
 
-export type TTSServer = 'openai' | 'edge' | 'microsoft';
+export type TTSServer = 'openai' | 'edge' | 'microsoft' | 'elevenlabs';
 
 export interface LobeAgentTTSConfig {
   showAllLocaleVoice?: boolean;
@@ -12,6 +13,7 @@ export interface LobeAgentTTSConfig {
   ttsService: TTSServer;
   voice: {
     edge?: string;
+    elevenlabs?: string;
     microsoft?: string;
     openai: string;
   };